requirements.txt
gradio>=4.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
python-pptx>=0.6.21
beautifulsoup4>=4.12.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0 